Move NavigationExtensions.kt to :app:ui:base
