+++ app/ui/legacy/build.gradle
@@ -27,8 +27,6 @@ dependencies {
     implementation "androidx.recyclerview:recyclerview:${versions.androidxRecyclerView}"
     implementation "androidx.lifecycle:lifecycle-viewmodel-ktx:${versions.androidxLifecycle}"
     implementation "androidx.lifecycle:lifecycle-livedata-ktx:${versions.androidxLifecycle}"
-    implementation "androidx.navigation:navigation-fragment-ktx:${versions.androidxNavigation}"
-    implementation "androidx.navigation:navigation-ui-ktx:${versions.androidxNavigation}"
     implementation "androidx.constraintlayout:constraintlayout:${versions.androidxConstraintLayout}"
     implementation "androidx.cardview:cardview:${versions.androidxCardView}"
     implementation "androidx.localbroadcastmanager:localbroadcastmanager:1.0.0"
