+++ app/ui/legacy/src/main/java/com/fsck/k9/ui/managefolders/ManageFoldersActivity.kt
@@ -10,7 +10,7 @@ import androidx.navigation.ui.setupActionBarWithNavController
 import com.fsck.k9.Account
 import com.fsck.k9.ui.R
 import com.fsck.k9.ui.base.K9Activity
-import com.fsck.k9.ui.findNavController
+import com.fsck.k9.ui.base.extensions.findNavController
 
 class ManageFoldersActivity : K9Activity() {
     private lateinit var navController: NavController
